Image solver and annotate image
